// backend/src/modules/task/core/model/Task.d.ts
import Entity from '../../../shared/core/model/Entity';
import DateTime from '../../../shared/core/objects/DateTime';
import None from '../../../shared/core/objects/None';
import TaskId from '../objects/TaskId';
import TaskTitle from '../objects/TaskTitle';
import TaskState from '../objects/TaskState';
import IdEntity from '../../../shared/core/objects/IdEntity';
import type TaskParams from '../interface/TaskParams';
import Text from '../../../shared/core/objects/Text';
import Collection from '../../../shared/core/objects/Collection';
import IntNumber from '../../../shared/core/objects/IntNumber';
export default class Task extends Entity {
    private title;
    private listContainer;
    private positionInList;
    private state;
    private archived;
    private description;
    private startDate;
    private dueDate;
    private isStarted;
    private isOverdue;
    private categories;
    private assigned;
    private constructor();
    static create(title: TaskTitle, listContainer: IdEntity, positionInList: IntNumber, state: TaskState, archived: boolean, id: TaskId, idProject: IdEntity, description: Text | None, startDate: DateTime | None, dueDate: DateTime | None, categories: Collection, assigned: Collection, actor: IdEntity, key: string): Task;
    static fromPrimitives(params: TaskParams): Task;
    delete(actor: IdEntity, key: string): void;
    removeCategory(category: IdEntity, actor: IdEntity, key: string): void;
    removeAssigned(assigned: IdEntity, actor: IdEntity, key: string): void;
    move(list: IdEntity, positionInList: IntNumber, actor: IdEntity, key: string): void;
    unarchive(actor: IdEntity, key: string): void;
    archive(actor: IdEntity, key: string): void;
    assignMember(actor: IdEntity, key: string, assigned: IdEntity): void;
    updateStartDate(date: DateTime, actor: IdEntity, key: string): void;
    updateDueDate(date: DateTime, actor: IdEntity, key: string): void;
    updateTitle(title: TaskTitle, actor: IdEntity, key: string): void;
    updateDescription(description: Text, actor: IdEntity, key: string): void;
    addCategory(category: IdEntity, actor: IdEntity, key: string): void;
    markAsFinished(actor: IdEntity, key: string): void;
    markAsPending(actor: IdEntity, key: string): void;
    protected isArchived(): boolean;
    protected isCompleted(): boolean;
    exportToProject(idProject: IdEntity, idList: IdEntity): void;
    overDue(): boolean;
    setStarted(key: string): void;
    setOverDue(key: string): void;
    getIdProject(): IdEntity;
    getTitle(): TaskTitle;
    getListContainer(): IdEntity;
    getState(): TaskState;
    getDescription(): Text | None;
    getStartDate(): DateTime | None;
    getDueDate(): DateTime | None;
    getCategories(): Collection;
    getAssigned(): Collection;
    getIsStarted(): boolean;
    getIsOverdue(): boolean;
    toPrimitives(): TaskParams;
}
//# sourceMappingURL=Task.d.ts.map